modified:   definitions.c 	modified:   exec_operators.c 	modified:   gen_operators.c 	modified:   other_functions.c 	modified:   stacks.c

+++ stacks.c
@@ -1,6 +1,7 @@
 #include "definitions.c"
 #include <stdio.h>
 #include <stdlib.h>
+#include <ctype.h>
 #include <complex.h>
 #include <string.h>
 
@@ -53,7 +54,7 @@ void *getnum_and_update_index(char *input, int *intptr)
 	{
 		exit(EXIT_FAILURE);
 	}
-	*intptr += abs((int)((long int)endptr - (long int)input));
+	*intptr += abs((int)((long long int)endptr - (long long int)input));
 	return vptr;
 }
